+++ src/applications/staticProps/GofAngle2.cpp
@@ -108,8 +108,8 @@ namespace OpenMD {
     RealType cosAngle2 = dot(dipole1, dipole2);
 
     RealType halfBin = (nAngleBins_ - 1) * 0.5;
-    int angleBin1 = halfBin * (cosAngle1 + 1.0);
-    int angleBin2 = halfBin * (cosAngle2 + 1.0);
+    int angleBin1 = int(halfBin * (cosAngle1 + 1.0));
+    int angleBin2 = int(halfBin * (cosAngle2 + 1.0));
 
     ++histogram_[angleBin1][angleBin2];    
     ++npairs_;
